separate eth input into its' own component
